backend/backend/requirements.txt
django
pyjwt
requests
